pyproject.toml
[project]
name = "s12p31b201"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "ultralytics>=8.3.130",
]
